+++ apps/dashboard/src/app/app-routing.module.ts
@@ -6,9 +6,11 @@ import {
   BaseComponent,
   DashboardLayoutComponent,
   EmptyComponent,
+  KnowledgeBoxSettingsComponent,
   PageNotFoundComponent,
   PageNotFoundModule,
   rootGuard,
+  SearchComponent,
   SelectAccountComponent,
   selectAccountKbGuard,
   SelectKbComponent,
@@ -23,12 +25,10 @@ import { RedirectComponent } from './redirect/redirect.component';
 import { AccountNUAComponent } from './account/account-nua/account-nua.component';
 import { NuaActivityComponent } from './account/account-nua/nua-activity/nua-activity.component';
 import { AccountUsersComponent } from './account/account-users/account-users.component';
-import { SearchComponent } from './search/search.component';
 import {
   KnowledgeBoxComponent,
   KnowledgeBoxHomeComponent,
   KnowledgeBoxKeysComponent,
-  KnowledgeBoxProfileComponent,
   KnowledgeBoxUsersComponent,
 } from './knowledge-box';
 import { inviteGuard } from './invite/invite.guard';
@@ -143,7 +143,7 @@ const routes: Routes = [
               },
               {
                 path: 'manage',
-                component: KnowledgeBoxProfileComponent,
+                component: KnowledgeBoxSettingsComponent,
               },
               {
                 path: 'training',
@@ -164,7 +164,8 @@ const routes: Routes = [
               },
               {
                 path: 'widget-generator',
-                loadChildren: () => import('./widgets/widgets.module').then((m) => m.WidgetsModule),
+                loadChildren: () =>
+                  import('../../../../libs/common/src/lib/widgets/widgets.module').then((m) => m.WidgetsModule),
               },
             ],
           },