+++ libs/common/src/lib/knowledge-box-settings/index.ts
@@ -1,0 +1,2 @@
+export * from './knowledge-box-settings.component';
+export * from './knowledge-box-settings.module';