+++ libs/common/src/lib/search/search.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SearchComponent } from './search.component';
-import { PaTogglesModule, PaTranslateModule } from '@guillotinaweb/pastanaga-angular';
+import { PaIconModule, PaTogglesModule, PaTranslateModule } from '@guillotinaweb/pastanaga-angular';
 
 @NgModule({
-  imports: [CommonModule, PaTogglesModule, PaTranslateModule],
+  imports: [CommonModule, PaTogglesModule, PaTranslateModule, PaIconModule],
   declarations: [SearchComponent],
   exports: [SearchComponent],
 })